+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project provides a central API and Dashboard for communication between appl
 - Adds middleware to check user auth token
 - Adds Users Sync from Dolibarr (FirstName, Lastname, Email, Password, Username, JSON Blob of Object)
 - Adds Projects Sync from Dolibarr
+- Adds Workorder model and  Sync from Dolibarr to Conduit
 ### Changes
 ### Removes
 ### Fixes